+++ CHANGELOG.md
@@ -23,14 +23,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 ## [Version 4] 2018-12-01
-
 - change all the sparse matrix structure
+- remove all map matrix jan 2019 
 
 ### Added
-- surface finite elemet (in progress)
+- surface finite element (in progress)
 
 - nElementonB (version 2d and 3d of nTonEge)
 - area  ( same the lenEdge) in 3d
+
+- add function labels to get the array of label of a  mesh 
+- add function regions to get the array of label of a  mesh 
+- correct big bug in 	toRarray,toZarray, toCarray  transform [ ... ] array to int[int], real[int], complex[int]  
 
 ## [Unreleased] 2018-08-31